refactor(psxed-project): extract test module and UI type cluster from lib.rs
First slice of splitting the 14.6k-line data-model lib.rs:
- ~2.8k-line #[cfg(test)] mod tests -> tests.rs
- UI authoring types (UiAnchor, UiRect, UiAction, UiSfx, OptionKind,
  UiFontChoice, UiImageEffect, UiNodeKind, UiNode, UiScene + serde
  helpers) -> ui_types.rs, re-exported via `pub use` so the
  psxed_project::Ui* public paths are unchanged.

lib.rs 14,664 -> 10,159 lines. 277 tests pass, whole editor workspace
builds (downstream consumers unaffected).
